[Incomplete] Re-add notify package
