+++ sdk/dynamicconfig/metric/controller/notify/basic/servicereader.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package push
+package basic
 
 import (
 	"bytes"
